LU-2592 tests: do not remove $TMP/*active in rpc.sh
The $TMP/*active files record the current active facets info
under failover test configuration. They are removed from
init_test_env() initially. If failover tests use do_rpc_nodes()
which performs rpc.sh and in which init_test_env() is used,
then the active facets info will be lost during the testing.

This patch fixes the above issue by introducing an RPC_MODE
variable which controls that the $TMP/*active files will not
be removed from init_test_env() in rpc.sh.

Test-Parameters: envdefinitions=SLOW=yes,ENABLE_QUOTA=yes \
clientcount=4 osscount=2 mdscount=2 austeroptions=-R \
failover=true useiscsi=true testlist=replay-vbr

Reviewed-on: http://review.whamcloud.com/5126
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>

+++ lustre/tests/test-framework.sh
@@ -113,6 +113,7 @@ init_test_env() {
     export TESTSUITE=`basename $0 .sh`
     export TEST_FAILED=false
     export FAIL_ON_SKIP_ENV=${FAIL_ON_SKIP_ENV:-false}
+    export RPC_MODE=${RPC_MODE:-false}
 
     export MKE2FS=${MKE2FS:-mke2fs}
     export DEBUGFS=${DEBUGFS:-debugfs}
@@ -247,7 +248,9 @@ init_test_env() {
     # print the durations of each test if "true"
     DDETAILS=${DDETAILS:-false}
     [ "$TESTSUITELOG" ] && rm -f $TESTSUITELOG || true
-    rm -f $TMP/*active
+    if ! $RPC_MODE; then
+        rm -f $TMP/*active
+    fi
 }
 
 case `uname -r` in
